@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.marconatalini.eurostep"
         minSdkVersion 21
-        targetSdkVersion 26
-        versionCode 6
-        versionName "autochiudilo"
+        targetSdkVersion 29
+        versionCode 7
+        versionName "symfony5powered"
     }
     buildTypes {
         release {
@@ -17,18 +17,24 @@ android {
     }
     productFlavors {
     }
+    compileOptions {
+        sourceCompatibility 1.8
+        targetCompatibility 1.8
+    }
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     //compile 'com.journeyapps:zxing-android-embedded:3.0.2@aar'
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    implementation 'com.android.support:design:26.1.0'
-    implementation 'com.android.support:recyclerview-v7:26.1.0'
-    implementation 'com.google.android.gms:play-services-appindexing:9.0.2'
-    implementation 'com.journeyapps:zxing-android-embedded:3.3.0'
-    implementation 'com.google.zxing:core:3.4.0'
-    implementation 'com.mcxiaoke.volley:library:1.0.19'
-    implementation 'com.android.support:support-v4:26.1.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.material:material:1.1.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+//    implementation 'com.google.android.gms:play-services-appindexing:9.8.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation 'junit:junit:4.12'
+    implementation 'com.android.volley:volley:1.1.1'
+
+    implementation('com.journeyapps:zxing-android-embedded:4.1.0') { transitive = false }
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.zxing:core:3.3.0'
 }
