@@ -2,9 +2,8 @@ package com.marconatalini.eurostep;
 
 import android.app.SearchManager;
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import android.widget.TextView;
@@ -15,12 +14,12 @@ import com.android.volley.Request;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
+import com.marconatalini.eurostep.tool.DateFromJsonTimestampString;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
@@ -34,8 +33,9 @@ public class DettOrdineLaActivity extends AppCompatActivity {
     TextView ordine_lotto, cliente, riferimento, scadenza, finitura,
             cornici, complementari, tagli, montaggi, lav_cnc, lamiere;
 
-    String json_url_dett = "http://" + MainActivity.WEBSERVER_IP + "/dettaglio_ordineLA.php";
-    String json_url_step = "http://" + MainActivity.WEBSERVER_IP + "/avanzamento_ordineLA.php";
+//    String json_url_dett = "http://" + MainActivity.WEBSERVER_IP + "/dettaglio_ordineLA.php";
+//    String json_url_step = "http://" + MainActivity.WEBSERVER_IP + "/avanzamento_ordineLA.php";
+
     ListView lista_step;
     ArrayList<String> record_step = new ArrayList<String>();
     ArrayAdapter<String> adapter;
@@ -103,33 +103,31 @@ public class DettOrdineLaActivity extends AppCompatActivity {
                 Lotto = extras.getString("lotto");
             }
 
+            String json_url_dett = "http://" + MainActivity.WEBSERVER_IP + "/dettaglio/" + Numero + "/" + Lotto;
+
             final String finalNumero = Numero;
             final String finalLotto = Lotto;
-            StringRequest jsonObjectRequest = new StringRequest(Request.Method.POST, json_url_dett,
+            StringRequest jsonObjectRequest = new StringRequest(Request.Method.GET, json_url_dett,
                     new Response.Listener<String>() {
                         @Override
                         public void onResponse(String response) {
                             try {
                                 JSONObject jsonObject = new JSONObject(response);
-                                ordine_lotto.setText(jsonObject.getString("numero") + "_" + jsonObject.getString("lotto"));
+
+                                ordine_lotto.setText(String.format("%s_%s", jsonObject.getString("numero"), jsonObject.getString("lotto")));
                                 cliente.setText(jsonObject.getString("xragsoc"));
-                                riferimento.setText("Rif. " + jsonObject.getString("riferimento"));
-                                SimpleDateFormat in_format = new SimpleDateFormat("yyyy-MM-dd", Locale.ITALIAN);
+                                riferimento.setText(String.format("Rif. %s", jsonObject.getString("riferimento")));
+//                                SimpleDateFormat in_format = new SimpleDateFormat("yyyy-MM-dd", Locale.ITALIAN);
                                 SimpleDateFormat out_format = new SimpleDateFormat("dd MMMM yyyy", Locale.ITALIAN);
-                                Date DataScadenza = null;
-                                try {
-                                    DataScadenza = in_format.parse(jsonObject.getString("data_scadenza_produzione").substring(0, 9));
-                                } catch (ParseException e) {
-                                    e.printStackTrace();
-                                }
+                                Date DataScadenza = new DateFromJsonTimestampString(jsonObject.getString("dataScadenzaProduzione")).getDate();
                                 scadenza.setText(out_format.format(DataScadenza));
                                 finitura.setText(jsonObject.getString("finitura"));
-                                cornici.setText(jsonObject.getString("n_cornici"));
-                                complementari.setText(jsonObject.getString("n_complementari"));
-                                tagli.setText(jsonObject.getString("n_tagli"));
-                                montaggi.setText(jsonObject.getString("n_montaggi"));
-                                lav_cnc.setText(jsonObject.getString("n_lavorazioni_CNC"));
-                                lamiere.setText(jsonObject.getString("n_lamiere"));
+                                cornici.setText(jsonObject.getString("nCornici"));
+                                complementari.setText(jsonObject.getString("nComplementari"));
+                                tagli.setText(jsonObject.getString("nTagli"));
+                                montaggi.setText(jsonObject.getString("nMontaggi"));
+                                lav_cnc.setText(jsonObject.getString("nLavorazioniCNC"));
+                                lamiere.setText(jsonObject.getString("nLamiere"));
 
                             } catch (JSONException e) {
                                 e.printStackTrace();
@@ -157,8 +155,9 @@ public class DettOrdineLaActivity extends AppCompatActivity {
                 }
             };
 
+            String json_url_step = "http://" + MainActivity.WEBSERVER_IP + "/avanzamento/" + Numero + "/" + Lotto;
 
-            StringRequest jsonArrayRequest = new StringRequest(Request.Method.POST, json_url_step,
+            StringRequest jsonArrayRequest = new StringRequest(Request.Method.GET, json_url_step,
                     new Response.Listener<String>() {
                         @Override
                         public void onResponse(String response) {
@@ -170,27 +169,25 @@ public class DettOrdineLaActivity extends AppCompatActivity {
                                 e.printStackTrace();
                             }
 
-                            while (count < jsonArray.length()) {
+                            while (jsonArray != null && count < jsonArray.length()) {
                                 try {
                                     JSONObject jsonObject = jsonArray.getJSONObject(count);
                                     SimpleDateFormat in_format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss", Locale.ITALIAN);
                                     SimpleDateFormat out_format = new SimpleDateFormat("dd/MM HH:mm", Locale.ITALIAN);
-                                    Date DataStep = null;
-                                    DataStep = in_format.parse(jsonObject.getString("timestamp"));
-                                    String bool_fine = jsonObject.getString("inizio_fine");
+                                    Date dataStep = new DateFromJsonTimestampString(jsonObject.getString("timestamp")).getDate();
+
+                                    boolean bool_fine = jsonObject.getBoolean("inizioFine");
                                     String stato = "";
-                                    if (bool_fine.equals("1")) {
+                                    if (bool_fine == true) {
                                         stato = "FINE";
                                     } else {
                                         stato = "inizio";
                                     }
 
-                                    record_step.add(out_format.format(DataStep) + " " + stato + " " + jsonObject.getString("descrizione") + " " + jsonObject.getString("nome_operatore"));
+                                    record_step.add(String.format("%s %s %s %s",out_format.format(dataStep), stato, jsonObject.getString("descrizione"), jsonObject.getString("operatore")));
                                     count++;
 
                                 } catch (JSONException e) {
-                                    e.printStackTrace();
-                                } catch (ParseException e) {
                                     e.printStackTrace();
                                 }
                             }
@@ -204,6 +201,7 @@ public class DettOrdineLaActivity extends AppCompatActivity {
                     error.printStackTrace();
                 }
             }) {
+/*
                 @Override
                 protected Map<String, String> getParams() throws AuthFailureError {
                     //Creating parameters
@@ -216,6 +214,7 @@ public class DettOrdineLaActivity extends AppCompatActivity {
                     //returning parameters
                     return params;
                 }
+*/
             };
 
             MySingleton.getInstance(DettOrdineLaActivity.this).addToRequestque(jsonObjectRequest);
